Updates to controls, jumpscare, digipen logo
1) had playtester feedback that Q is too close to controls so accidentally exit. hence all exit is now changed to ESCAPE.

^additionally for exiting to mainmenu (in win page) the Q changed to M

2) changed IsAreaClick float to s32 for mouse variables to remove warning of argument conversion

3) ensure digipen logo is playing for at least 2 seconds before skipping (for now is still 0 so its not annoying while we code and debug)

4) updated jumpscare with random image

+++ Source File/Collision.cpp
@@ -2,7 +2,7 @@
 
 #include "Collision.h"
 
-bool IsAreaClicked(float area_center_x, float area_center_y, float area_width, float area_height, float click_x, float click_y) {
+bool IsAreaClicked(float area_center_x, float area_center_y, float area_width, float area_height, s32 click_x, s32 click_y) {
     
     // Check if click is within the area calculated from the center
     if (click_x >= area_center_x - area_width / 2 &&
