+++ Source File/Collision.h
@@ -5,5 +5,5 @@
 // Sharon: feel free to add any collision that you all neeed in here
 
 // Collision between cursor and rectangle
-bool IsAreaClicked(float area_center_x, float area_center_y, float area_width, float area_height, float click_x, float click_y);
+bool IsAreaClicked(float area_center_x, float area_center_y, float area_width, float area_height, s32 click_x, s32 click_y);
